+++ components/pm-subscription-handler/Changelog.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Updated to get NFs list when requesting a specific subscription (DCAEGEN2-2992)
 * AAI Event handler changes with new subscription format (DCAEGEN2-2912)
 * Read NFS associated with MG by using MGName and subName(DCAEGEN2-2993)
+* Delete subscription API by Name(DCAEGEN2-2821)
 
 ## [1.3.2]
 ### Changed